--- components/NavBarItems.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-
-const NavBarItems = (props) => {
-  return (
-    <div className="nav-items">
-      <ul>
-        <li>Home </li>
-        <li>About Us</li>
-        <li>Contact</li>
-        <li>Cart</li>
-      </ul>
-    </div>
-  );
-};
-
-export default NavBarItems;
