+++ config.js
@@ -1,0 +1,822 @@
+const restaurantList = [
+  {
+    type: "restaurant",
+    data: {
+      type: "F",
+      id: "528813",
+      name: "Bakingo",
+      uuid: "7253c33d-6733-4abf-8343-e5d16d424e3f",
+      city: "12",
+      area: "Mansarovar",
+      totalRatingsString: "1000+ ratings",
+      cloudinaryImageId: "tmhaokxkj2uzqnfq8b4v",
+      cuisines: ["Bakery", "Desserts", "Beverages", "Snacks"],
+      tags: [],
+      costForTwo: 29900,
+      costForTwoString: "₹299 FOR TWO",
+      deliveryTime: 28,
+      minDeliveryTime: 28,
+      maxDeliveryTime: 28,
+      slaString: "28 MINS",
+      lastMileTravel: 5.5,
+      slugs: {
+        restaurant: "bakingo-jaipur-manasarovar",
+        city: "jaipur",
+      },
+      cityState: "12",
+      address:
+        "111, 395C, Shipra Path, Jhalana Chhod, Mansarovar, Jaipur, Rajasthan 302020",
+      locality: "Shipra Path",
+      parentId: 3818,
+      unserviceable: false,
+      veg: false,
+      select: false,
+      favorite: false,
+      tradeCampaignHeaders: [],
+      ribbon: [
+        {
+          type: "PROMOTED",
+        },
+      ],
+      chain: [],
+      feeDetails: {
+        fees: [
+          {
+            name: "distance",
+            fee: 5200,
+            message: "",
+          },
+          {
+            name: "time",
+            fee: 0,
+            message: "",
+          },
+          {
+            name: "special",
+            fee: 0,
+            message: "",
+          },
+        ],
+        totalFees: 5200,
+        message: "",
+        title: "Delivery Charge",
+        amount: "5200",
+        icon: "",
+      },
+      availability: {
+        opened: true,
+        nextOpenMessage: "",
+        nextCloseMessage: "",
+      },
+      longDistanceEnabled: 0,
+      rainMode: "NONE",
+      thirdPartyAddress: false,
+      thirdPartyVendor: "",
+      adTrackingID: "cid=6509472~p=1~eid=00000187-95b5-68e8-2518-67f800b10149",
+      badges: {
+        imageBased: [],
+        textBased: [],
+        textExtendedBadges: [],
+      },
+      lastMileTravelString: "5.5 kms",
+      hasSurge: false,
+      sla: {
+        restaurantId: "528813",
+        deliveryTime: 28,
+        minDeliveryTime: 28,
+        maxDeliveryTime: 28,
+        lastMileTravel: 5.5,
+        lastMileDistance: 0,
+        serviceability: "SERVICEABLE",
+        rainMode: "NONE",
+        longDistance: "NOT_LONG_DISTANCE",
+        preferentialService: false,
+        iconType: "EMPTY",
+      },
+      promoted: true,
+      avgRating: "4.4",
+      totalRatings: 1000,
+      new: false,
+    },
+    subtype: "basic",
+  },
+  {
+    type: "restaurant",
+    data: {
+      type: "F",
+      id: "655295",
+      name: "Siddhi Vinayak Restaurant",
+      uuid: "26ed9351-7a23-49a1-9e41-41f1a0f54c0b",
+      city: "12",
+      area: "Sanganer",
+      totalRatingsString: "50+ ratings",
+      cloudinaryImageId: "a533a43b8677a5582c28bb44eb30adbf",
+      cuisines: ["North Indian", "Indian", "Biryani"],
+      tags: [],
+      costForTwo: 21000,
+      costForTwoString: "₹210 FOR TWO",
+      deliveryTime: 30,
+      minDeliveryTime: 30,
+      maxDeliveryTime: 30,
+      slaString: "30 MINS",
+      lastMileTravel: 1.899999976158142,
+      slugs: {
+        restaurant:
+          "santa-banta-veg.-restro-mansarovar-extension-mansarovar-extension",
+        city: "jaipur",
+      },
+      cityState: "12",
+      address: "Phot 3 Shyam Nager 3rd, Diggi Road Sanganer,Jaipur 302029",
+      locality: "Diggi Road",
+      parentId: 187779,
+      unserviceable: false,
+      veg: true,
+      select: false,
+      favorite: false,
+      tradeCampaignHeaders: [],
+      aggregatedDiscountInfo: {
+        header: "50% off",
+        shortDescriptionList: [
+          {
+            meta: "50% off | Use TRYNEW",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+        ],
+        descriptionList: [
+          {
+            meta: "50% off up to ₹100 | Use code TRYNEW",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+        ],
+        subHeader: "",
+        headerType: 0,
+        superFreedel: "",
+      },
+      aggregatedDiscountInfoV2: {
+        header: "50% OFF",
+        shortDescriptionList: [
+          {
+            meta: "Use TRYNEW",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+        ],
+        descriptionList: [
+          {
+            meta: "50% off up to ₹100 | Use code TRYNEW",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+        ],
+        subHeader: "",
+        headerType: 0,
+        superFreedel: "",
+      },
+      chain: [],
+      feeDetails: {
+        fees: [
+          {
+            name: "distance",
+            fee: 2800,
+            message: "",
+          },
+          {
+            name: "time",
+            fee: 0,
+            message: "",
+          },
+          {
+            name: "special",
+            fee: 0,
+            message: "",
+          },
+        ],
+        totalFees: 2800,
+        message: "",
+        title: "Delivery Charge",
+        amount: "2800",
+        icon: "",
+      },
+      availability: {
+        opened: true,
+        nextOpenMessage: "",
+        nextCloseMessage: "",
+      },
+      longDistanceEnabled: 0,
+      rainMode: "NONE",
+      thirdPartyAddress: false,
+      thirdPartyVendor: "",
+      adTrackingID: "",
+      badges: {
+        imageBased: [],
+        textBased: [],
+        textExtendedBadges: [],
+      },
+      lastMileTravelString: "1.8 kms",
+      hasSurge: false,
+      sla: {
+        restaurantId: "655295",
+        deliveryTime: 30,
+        minDeliveryTime: 30,
+        maxDeliveryTime: 30,
+        lastMileTravel: 1.899999976158142,
+        lastMileDistance: 0,
+        serviceability: "SERVICEABLE",
+        rainMode: "NONE",
+        longDistance: "NOT_LONG_DISTANCE",
+        preferentialService: false,
+        iconType: "EMPTY",
+      },
+      promoted: false,
+      avgRating: "3.9",
+      totalRatings: 50,
+      new: false,
+    },
+    subtype: "basic",
+  },
+  {
+    type: "restaurant",
+    data: {
+      type: "F",
+      id: "579415",
+      name: "Magic Moments",
+      uuid: "12255d40-bf48-4fee-96d0-9e6107a75aec",
+      city: "12",
+      area: "Pratap Nagar",
+      totalRatingsString: "Too Few Ratings",
+      cloudinaryImageId: "xkqn0ngqvmg7jnr48zvy",
+      cuisines: ["Italian-American", "Chinese", "Fast Food"],
+      tags: [],
+      costForTwo: 20000,
+      costForTwoString: "₹200 FOR TWO",
+      deliveryTime: 29,
+      minDeliveryTime: 29,
+      maxDeliveryTime: 29,
+      slaString: "29 MINS",
+      lastMileTravel: 1.899999976158142,
+      slugs: {
+        restaurant: "magic-moments-mansarovar-extension-mansarovar-extension",
+        city: "jaipur",
+      },
+      cityState: "12",
+      address:
+        "shop number 4, shyam nagar, tiwary road behind of diggi malpura rd, sanganer, Manasover jaipur rajasthan ,,Jaipur-II Ward 34,Jaipur,Rajasthan-302020",
+      locality: "Sanganer",
+      parentId: 129344,
+      unserviceable: false,
+      veg: false,
+      select: false,
+      favorite: false,
+      tradeCampaignHeaders: [],
+      aggregatedDiscountInfo: {
+        header: "FLAT250 off",
+        shortDescriptionList: [
+          {
+            meta: "FLAT250 off | Use FLATDEAL",
+            discountType: "Flat",
+            operationType: "RESTAURANT",
+          },
+        ],
+        descriptionList: [
+          {
+            meta: "FLAT250 off | Use FLATDEAL",
+            discountType: "Flat",
+            operationType: "RESTAURANT",
+          },
+        ],
+        subHeader: "",
+        headerType: 0,
+        superFreedel: "",
+      },
+      aggregatedDiscountInfoV2: {
+        header: "₹250 OFF",
+        shortDescriptionList: [
+          {
+            meta: "Use FLATDEAL",
+            discountType: "Flat",
+            operationType: "RESTAURANT",
+          },
+        ],
+        descriptionList: [
+          {
+            meta: "FLAT250 off | Use FLATDEAL",
+            discountType: "Flat",
+            operationType: "RESTAURANT",
+          },
+        ],
+        subHeader: "",
+        headerType: 0,
+        superFreedel: "",
+      },
+      chain: [],
+      feeDetails: {
+        fees: [
+          {
+            name: "distance",
+            fee: 2800,
+            message: "",
+          },
+          {
+            name: "time",
+            fee: 0,
+            message: "",
+          },
+          {
+            name: "special",
+            fee: 0,
+            message: "",
+          },
+        ],
+        totalFees: 2800,
+        message: "",
+        title: "Delivery Charge",
+        amount: "2800",
+        icon: "",
+      },
+      availability: {
+        opened: true,
+        nextOpenMessage: "",
+        nextCloseMessage: "",
+      },
+      longDistanceEnabled: 0,
+      rainMode: "NONE",
+      thirdPartyAddress: false,
+      thirdPartyVendor: "",
+      adTrackingID: "",
+      badges: {
+        imageBased: [],
+        textBased: [],
+        textExtendedBadges: [],
+      },
+      lastMileTravelString: "1.8 kms",
+      hasSurge: false,
+      sla: {
+        restaurantId: "579415",
+        deliveryTime: 29,
+        minDeliveryTime: 29,
+        maxDeliveryTime: 29,
+        lastMileTravel: 1.899999976158142,
+        lastMileDistance: 0,
+        serviceability: "SERVICEABLE",
+        rainMode: "NONE",
+        longDistance: "NOT_LONG_DISTANCE",
+        preferentialService: false,
+        iconType: "EMPTY",
+      },
+      promoted: false,
+      avgRating: "--",
+      totalRatings: 0,
+      new: false,
+    },
+    subtype: "basic",
+  },
+  {
+    type: "restaurant",
+    data: {
+      type: "F",
+      id: "558588",
+      name: "Rominus Pizza & Burger",
+      uuid: "ebd186ee-0b04-4913-9ec3-fcfc33c8f78e",
+      city: "12",
+      area: "Pratap Nagar",
+      totalRatingsString: "1000+ ratings",
+      cloudinaryImageId: "9ec9ffd900c24ef751e2f34ba3d2fd4b",
+      cuisines: [
+        "Pizzas",
+        "Italian-American",
+        "Italian",
+        "American",
+        "Desserts",
+        "Beverages",
+        "Fast Food",
+        "Snacks",
+        "Continental",
+        "Bakery",
+      ],
+      tags: [],
+      costForTwo: 20000,
+      costForTwoString: "₹200 FOR TWO",
+      deliveryTime: 34,
+      minDeliveryTime: 34,
+      maxDeliveryTime: 34,
+      slaString: "34 MINS",
+      lastMileTravel: 6.199999809265137,
+      slugs: {
+        restaurant: "rominus-pizza-and-burger-pratap-nagar-pratap-nagar",
+        city: "jaipur",
+      },
+      cityState: "12",
+      address:
+        "SHOP NO. LGF-02-03 DA RISE NEAR D-MART TONK ROAD PRATAP NAGAR JAIPUR, Sanganer, Jaipur, Rajasthan-302033",
+      locality: "Tonk Road",
+      parentId: 8387,
+      unserviceable: false,
+      veg: false,
+      select: false,
+      favorite: false,
+      tradeCampaignHeaders: [],
+      aggregatedDiscountInfo: {
+        header: "50% off",
+        shortDescriptionList: [
+          {
+            meta: "50% off | Use TRYNEW",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+        ],
+        descriptionList: [
+          {
+            meta: "50% off up to ₹100 | Use code TRYNEW",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+        ],
+        subHeader: "",
+        headerType: 0,
+        superFreedel: "",
+      },
+      aggregatedDiscountInfoV2: {
+        header: "50% OFF",
+        shortDescriptionList: [
+          {
+            meta: "Use TRYNEW",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+        ],
+        descriptionList: [
+          {
+            meta: "50% off up to ₹100 | Use code TRYNEW",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+        ],
+        subHeader: "",
+        headerType: 0,
+        superFreedel: "",
+      },
+      ribbon: [
+        {
+          type: "PROMOTED",
+        },
+      ],
+      chain: [],
+      feeDetails: {
+        fees: [
+          {
+            name: "time",
+            fee: 0,
+            message: "",
+          },
+          {
+            name: "distance",
+            fee: 6200,
+            message: "",
+          },
+          {
+            name: "special",
+            fee: 0,
+            message: "",
+          },
+        ],
+        totalFees: 6200,
+        message: "",
+        title: "Delivery Charge",
+        amount: "6200",
+        icon: "",
+      },
+      availability: {
+        opened: true,
+        nextOpenMessage: "",
+        nextCloseMessage: "",
+      },
+      longDistanceEnabled: 0,
+      rainMode: "NONE",
+      thirdPartyAddress: false,
+      thirdPartyVendor: "",
+      adTrackingID: "cid=6518550~p=4~eid=00000187-95b5-68e8-2518-67f900b10445",
+      badges: {
+        imageBased: [],
+        textBased: [],
+        textExtendedBadges: [],
+      },
+      lastMileTravelString: "6.1 kms",
+      hasSurge: false,
+      sla: {
+        restaurantId: "558588",
+        deliveryTime: 34,
+        minDeliveryTime: 34,
+        maxDeliveryTime: 34,
+        lastMileTravel: 6.199999809265137,
+        lastMileDistance: 0,
+        serviceability: "SERVICEABLE",
+        rainMode: "NONE",
+        longDistance: "NOT_LONG_DISTANCE",
+        preferentialService: false,
+        iconType: "EMPTY",
+      },
+      promoted: true,
+      avgRating: "3.8",
+      totalRatings: 1000,
+      new: false,
+    },
+    subtype: "basic",
+  },
+  {
+    type: "restaurant",
+    data: {
+      type: "F",
+      id: "583473",
+      name: "Wow !! Momes",
+      uuid: "e408bfcc-6727-4bcd-804d-a7a4248a59d2",
+      city: "12",
+      area: "Pratap Nagar",
+      totalRatingsString: "Too Few Ratings",
+      cloudinaryImageId: "xrlkomvztqvogtwmr0ch",
+      cuisines: ["Chinese"],
+      tags: [],
+      costForTwo: 20000,
+      costForTwoString: "₹200 FOR TWO",
+      deliveryTime: 26,
+      minDeliveryTime: 26,
+      maxDeliveryTime: 26,
+      slaString: "26 MINS",
+      lastMileTravel: 1.899999976158142,
+      slugs: {
+        restaurant: "wow-!!-momes-pratap-nagar-pratap-nagar",
+        city: "jaipur",
+      },
+      cityState: "12",
+      address:
+        "shop number 4, shyam nagar, tiwary road behind of diggi malpura rd, sanganer, Manasover jaipur rajasthan ,,Jaipur-II Ward 34,Jaipur,Rajasthan-302020",
+      locality: "Sanganer",
+      parentId: 349225,
+      unserviceable: false,
+      veg: false,
+      select: false,
+      favorite: false,
+      tradeCampaignHeaders: [],
+      aggregatedDiscountInfo: {
+        header: "FLAT100 off",
+        shortDescriptionList: [
+          {
+            meta: "FLAT100 off | Use FLATDEAL",
+            discountType: "Flat",
+            operationType: "RESTAURANT",
+          },
+        ],
+        descriptionList: [
+          {
+            meta: "FLAT100 off | Use FLATDEAL",
+            discountType: "Flat",
+            operationType: "RESTAURANT",
+          },
+        ],
+        subHeader: "",
+        headerType: 0,
+        superFreedel: "",
+      },
+      aggregatedDiscountInfoV2: {
+        header: "₹100 OFF",
+        shortDescriptionList: [
+          {
+            meta: "Use FLATDEAL",
+            discountType: "Flat",
+            operationType: "RESTAURANT",
+          },
+        ],
+        descriptionList: [
+          {
+            meta: "FLAT100 off | Use FLATDEAL",
+            discountType: "Flat",
+            operationType: "RESTAURANT",
+          },
+        ],
+        subHeader: "",
+        headerType: 0,
+        superFreedel: "",
+      },
+      chain: [],
+      feeDetails: {
+        fees: [
+          {
+            name: "distance",
+            fee: 2800,
+            message: "",
+          },
+          {
+            name: "time",
+            fee: 0,
+            message: "",
+          },
+          {
+            name: "special",
+            fee: 0,
+            message: "",
+          },
+        ],
+        totalFees: 2800,
+        message: "",
+        title: "Delivery Charge",
+        amount: "2800",
+        icon: "",
+      },
+      availability: {
+        opened: true,
+        nextOpenMessage: "",
+        nextCloseMessage: "",
+      },
+      longDistanceEnabled: 0,
+      rainMode: "NONE",
+      thirdPartyAddress: false,
+      thirdPartyVendor: "",
+      adTrackingID: "",
+      badges: {
+        imageBased: [],
+        textBased: [],
+        textExtendedBadges: [],
+      },
+      lastMileTravelString: "1.8 kms",
+      hasSurge: false,
+      sla: {
+        restaurantId: "583473",
+        deliveryTime: 26,
+        minDeliveryTime: 26,
+        maxDeliveryTime: 26,
+        lastMileTravel: 1.899999976158142,
+        lastMileDistance: 0,
+        serviceability: "SERVICEABLE",
+        rainMode: "NONE",
+        longDistance: "NOT_LONG_DISTANCE",
+        preferentialService: false,
+        iconType: "EMPTY",
+      },
+      promoted: false,
+      avgRating: "--",
+      totalRatings: 0,
+      new: false,
+    },
+    subtype: "basic",
+  },
+  {
+    type: "restaurant",
+    data: {
+      type: "F",
+      id: "67293",
+      name: "Sahni Pure Veg",
+      uuid: "e39c5ec5-e7e1-4f5d-aa66-02df79e280ec",
+      city: "12",
+      area: "Malviya Nagar",
+      totalRatingsString: "5000+ ratings",
+      cloudinaryImageId: "kzemkmy7vf91lnjdr99x",
+      cuisines: ["North Indian", "Biryani", "Punjabi", "Thalis", "Chinese"],
+      tags: [],
+      costForTwo: 30000,
+      costForTwoString: "₹300 FOR TWO",
+      deliveryTime: 33,
+      minDeliveryTime: 33,
+      maxDeliveryTime: 33,
+      slaString: "33 MINS",
+      lastMileTravel: 7.800000190734863,
+      slugs: {
+        restaurant: "sahni-pure-veg-model-town-malviya-nagar",
+        city: "jaipur",
+      },
+      cityState: "12",
+      address:
+        "2/14, Girdhar Marg, Sector 12, Malviya Nagar, Jaipur, Rajasthan",
+      locality: "Sector 12",
+      parentId: 174755,
+      unserviceable: false,
+      veg: false,
+      select: false,
+      favorite: false,
+      tradeCampaignHeaders: [],
+      aggregatedDiscountInfo: {
+        header: "70% off",
+        shortDescriptionList: [
+          {
+            meta: "70% off on all orders",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+          {
+            meta: "Free Jeera Chach on orders above ₹949",
+            discountType: "Freebie",
+            operationType: "RESTAURANT",
+          },
+        ],
+        descriptionList: [
+          {
+            meta: "70% off on all orders",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+          {
+            meta: "Free Jeera Chach on orders above ₹949",
+            discountType: "Freebie",
+            operationType: "RESTAURANT",
+          },
+        ],
+        subHeader: "",
+        headerType: 0,
+        superFreedel: "",
+      },
+      aggregatedDiscountInfoV2: {
+        header: "70% OFF",
+        shortDescriptionList: [
+          {
+            meta: "",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+        ],
+        descriptionList: [
+          {
+            meta: "70% off on all orders",
+            discountType: "Percentage",
+            operationType: "RESTAURANT",
+          },
+          {
+            meta: "Free Jeera Chach on orders above ₹949",
+            discountType: "Freebie",
+            operationType: "RESTAURANT",
+          },
+        ],
+        subHeader: "",
+        headerType: 0,
+        superFreedel: "",
+      },
+      ribbon: [
+        {
+          type: "PROMOTED",
+        },
+      ],
+      chain: [],
+      feeDetails: {
+        fees: [
+          {
+            name: "distance",
+            fee: 7200,
+            message: "",
+          },
+          {
+            name: "time",
+            fee: 0,
+            message: "",
+          },
+          {
+            name: "special",
+            fee: 0,
+            message: "",
+          },
+        ],
+        totalFees: 7200,
+        message: "",
+        title: "Delivery Charge",
+        amount: "7200",
+        icon: "",
+      },
+      availability: {
+        opened: true,
+        nextOpenMessage: "",
+        nextCloseMessage: "",
+      },
+      longDistanceEnabled: 0,
+      rainMode: "NONE",
+      thirdPartyAddress: false,
+      thirdPartyVendor: "",
+      adTrackingID: "cid=6517611~p=7~eid=00000187-95b5-68e8-2518-67fa00b1070f",
+      badges: {
+        imageBased: [],
+        textBased: [],
+        textExtendedBadges: [],
+      },
+      lastMileTravelString: "7.8 kms",
+      hasSurge: false,
+      sla: {
+        restaurantId: "67293",
+        deliveryTime: 33,
+        minDeliveryTime: 33,
+        maxDeliveryTime: 33,
+        lastMileTravel: 7.800000190734863,
+        lastMileDistance: 0,
+        serviceability: "SERVICEABLE",
+        rainMode: "NONE",
+        longDistance: "NOT_LONG_DISTANCE",
+        preferentialService: false,
+        iconType: "EMPTY",
+      },
+      promoted: true,
+      avgRating: "3.7",
+      totalRatings: 5000,
+      new: false,
+    },
+    subtype: "basic",
+  },
+];
+export const ImageId =
+  "https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_508,h_320,c_fill/";
+export default restaurantList;
