+++ index.html
@@ -9,6 +9,6 @@
   </head>
   <body>
     <div id="root">Not Rendered</div>
-    <script type="module" src="./App.js"></script>
+    <script type="module" src="./src/App.js"></script>
   </body>
 </html>
